[fx] move meta registration
